pyproject.toml
[project]
name = "py-microservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.35.0",
    "grpcio>=1.74.0",
    "grpcio-health-checking>=1.74.0",
    "grpcio-tools>=1.74.0",
    "injector>=0.22.0",
    "jinja2>=3.1.6",
    "json-repair>=0.51.0",
    "ortools>=9.14.6206",
    "pydantic-ai>=0.4.11",
    "python-dotenv>=1.1.1",
    "weasyprint>=66.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "factory-boy>=3.3.3",
    "flak>=0.1.5",
    "grpcio-testing>=1.74.0",
    "isort>=6.0.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.12.7",
]
